moving privacy account core=> main->privacy
