--- src/main/core/Resources/modules/account/privacy/store/index.js
@@ -1,8 +1,0 @@
-
-import {actions} from '#/main/core/account/privacy/store/actions'
-import {selectors} from '#/main/core/account/privacy/store/selectors'
-
-export {
-  actions,
-  selectors
-}
