+++ src/main/privacy/Resources/modules/account/privacy/components/main.jsx
@@ -27,7 +27,7 @@ const PrivacyMain = (props) =>
     title={trans('privacy')}
   >
     <ContentTitle
-      title={trans('terms_of_service')}
+      title={trans('terms_of_service',{}, 'terms_of_service')}
       style={{marginTop: 60}}
     />
