+++ src/main/privacy/Resources/modules/account/privacy/store/index.js
@@ -1,0 +1,8 @@
+
+import {actions} from '#/main/privacy/account/privacy/store/actions'
+import {selectors} from '#/main/privacy/account/privacy/store/selectors'
+
+export {
+  actions,
+  selectors
+}
