+++ src/main/privacy/Resources/translations/privacy.fr.json
@@ -1,8 +1,9 @@
 {
   "accept_terms": "Vous avez accepté les conditions d'utilisation de la plateforme.",
   "no_accept_terms": "Vous n'avez pas encore accepté les conditions d'utilisation de la plateforme.",
-  "title_my_datas": "Mes données personnelles",
-  "export_datas": "Exporter mes données",
+  "title_my_data": "Mes données personnelles",
+  "export_data": "Exporter mes données",
+  "country_storage": "Pays de stockage des données",
   "request_deletion" : "Demander la suppression de mon compte",
   "title_dialog_delete_account" : "Demande de suppression de mon compte",
   "message_dialog_delete_account" : "Êtes-vous sûr de vouloir demander la suppression de votre compte ?",
