+++ src/main/privacy/Resources/translations/terms_of_service.en.json
@@ -1,0 +1,7 @@
+{
+  "terms_of_service": "Terms of service",
+  "terms_of_service_activation_message": "Activate the terms of service",
+  "terms_of_service_activation_help": "New users will need to accept the terms to be able to use the platform.",
+  "terms_enabled_but_empty": "Terms of service can only be enabled if a text is provided in at least one language.",
+  "terms_of_service_acceptance": "I accept the terms of service"
+}
